fix(models-service): rename src/models/ → src/model_registry/
The previous fix anchored the .gitignore patterns to the runtime weight
cache locations, but that's a fragile workaround — anyone adding a new
`models/` directory anywhere in the tree could hit the same trap.

Renamed the source dir to remove the collision entirely. `model_registry`
also reads more accurately: the package is the registry/loader/downloader
layer, distinct from the runtime weight cache it manages.

Reverted the .gitignore tweaks since the bare `models/` pattern is now
unambiguous again (only the cache dir matches it).

Updated all imports in src/inference/, src/server.py, and the cross-
package references inside model_registry/ itself, plus the three doc-
comment path references that mentioned `models/`.
