+++ genai-engine/models-service/src/inference/claim_filter/classifier.py
@@ -33,7 +33,7 @@ import logging
 import numpy as np
 from opentelemetry import trace
 
-from models import loader
+from model_registry import loader
 from schemas import (
     ClaimClassification,
     ClaimClassifierResultEnum,
@@ -53,7 +53,7 @@ def classify(req: ClaimFilterRequest) -> ClaimFilterResponse:
     with tracer.start_as_current_span("claim_filter: classify"):
         result = classifier(req.texts)
 
-    # `result` from Classifier.forward (see models/classifier_arch.py): a dict
+    # `result` from Classifier.forward (see model_registry/classifier_arch.py): a dict
     # with keys "label" (np int array), "logit", "prob" (softmax distribution),
     # "pred_label_str" (list[str] from inv_label_map).
     labels: list[str] = result["pred_label_str"]
