+++ genai-engine/models-service/src/model_registry/classifier_arch.py
